reorganizing plus small docs updates
