+++ cleanlab_studio/internal/api/api.py
@@ -21,7 +21,7 @@ from cleanlab_studio.internal.types import JSONDict
 from cleanlab_studio.version import __version__
 
 
-base_url = os.environ.get("CLEANLAB_API_BASE_URL", "https://api.cleanlab.ai/api")
+base_url = os.environ.get("CLEANLAB_API_BASE_URL", "https://api.dev-bc26qf4m.cleanlab.ai/api")
 cli_base_url = f"{base_url}/cli/v0"
 upload_base_url = f"{base_url}/upload/v0"
 dataset_base_url = f"{base_url}/datasets"
